fix: avoid duplicative code & validate for specific RegExp on form change to disable it
