+++ src/utils/validateSnakeCase.ts
@@ -1,4 +1,4 @@
 export const validateSnakeCase = (value: string) => {
-  const snakeCaseRegex = /^[a-z0-9]+(_[a-z0-9]+)*$/;
+  const snakeCaseRegex = /^^[a-zA-Z0-9_-]{1,64}$/;
   return snakeCaseRegex.test(value);
 };
